Move RetryPolicy to a separate file
